Make releaseSomeActivities() a oneway call.
This is called from a GC handler hook in BinderInternal, from the
finalizer thread. It's a call from an app process into system_server. On
some devices, we observed this call taking a long time, causing
TimeoutExceptions on the finalizer thread.

Since this work is not critical, and when the GC runs is anyway
unpredictable, make releaseSomeActivities() oneway instead.

Bug: 118997212
Test: TH
Change-Id: I6b06917493a09a2fba63502c4bd1a203c184a62c
Merged-In: I6b06917493a09a2fba63502c4bd1a203c184a62c
Signed-off-by: Akash <akashniki@gmail.com>
Signed-off-by: Dhruv <dhruvgera61@gmail.com>